Move IOrderGenerator to OpenRA.Orders namespace.
