+++ OpenRA.Mods.Common/Orders/OrderGenerator.cs
@@ -12,6 +12,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Graphics;
+using OpenRA.Orders;
 
 namespace OpenRA.Mods.Common.Orders
 {
