+++ OpenRA.Mods.Common/Traits/SupportPowers/SelectDirectionalTarget.cs
@@ -14,6 +14,7 @@ using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Graphics;
 using OpenRA.Mods.Common.Widgets;
+using OpenRA.Orders;
 using OpenRA.Traits;
 using OpenRA.Widgets;
